Change how test requirements are listed
The pulp-ansible plugin distributes unit and functional tests for its
code alongside the code itself. Given this, the plugin should tell users
what they need to install in order to run those tests.

Define a new optional dependency called "test." It can be used like so:

    pip install pulp-ansible[test]

This install not only pulp-ansible and its hard dependencies, but also
everything needed for running unit tests and functional tests. The
tooling for actually *running* these tests once the dependences are
installed is not so great. This can be improved with some hard thought
and elbow grease.

See:
https://setuptools.readthedocs.io/en/latest/setuptools.html#declaring-extras-optional-features-with-their-own-dependencies