Travis pipeline needs to `make html`
The pipeline wasn't running `make html` correctly so it wouldn't fail if
there were docs errors. This causes the pipeline to run `make html`
correctly.

Also it adds in a doc_requirements.txt file which the pipeline uses to
install docs requirements.

It also fixes the docs verison number to be correct. We really need to
start using bumpversion.

Also it adds in the latest api schema which was forgotten too.

https://pulp.plan.io/issues/4571
closes #4571

asdf