Add is_highest field to CollectionVersion model
`is_highest` is a boolean field that indicates if a specific
collection version has the highest version in a set of available
collection versions.
This patch adds the field, migration and data migration for it.

NOTE: Since the semver comparison is implemented in Python, data
migration fetches all required data into memory. Even considering
limited number of fields that queries fetch, the operation may have
impact on performance or even fail if out of memory.

https://pulp.plan.io/issues/5223
closes #5223