Revert "add new upload serializer"
This reverts commit acc5c3ebdb1d7d901f271614305d7e88386504ef.

The Automation Hub project relied on the current behaviors of V3 API
adn this broke them. We do not have functional test coverage for
pulp_ansible in that area yet.

We hope to receive functional tests from them soon, which will allow us
to port onto the SingleContentArtifactUploadViewset and corresponding
serializer with confidence then.

https://pulp.plan.io/issues/5555
closes #5555