Fixes a broken test
A test was broken due to increased validation added to the
`galaxy-importer` dependnecy. Although that validation hasn't been
released, it was causing our tests to fail which use the master branch
of `galaxy-importer`.

This switches the collection being uploaded in the test to one which
passes the `galaxy-importer` validation requirements.

closes #8882