Fix constraints violation on is_highest in import
This field is not correct anyway and it is removed in newer versions. We
only set it to False here hoping it will not be worse but also as the
minimal possible change.

This is a backportable version of the fix.

relates to #1986

(cherry picked from commit 0474b113e45645b732eabca2e9f6a4ec4dc20a42)